fido2: Move to authentication dir
- Add authentication directory
- Move fido2 into authentication
