+++ subsys/CMakeLists.txt
@@ -12,6 +12,7 @@ add_subdirectory_ifdef(CONFIG_LORAWAN lorawan)
 add_compile_options($<TARGET_PROPERTY:compiler,warning_shadow_variables>)
 
 # zephyr-keep-sorted-start
+add_subdirectory(authentication)
 add_subdirectory(canbus)
 add_subdirectory(debug)
 add_subdirectory(fs)
@@ -44,7 +45,6 @@ add_subdirectory_ifdef(CONFIG_DAP dap)
 add_subdirectory_ifdef(CONFIG_DISK_ACCESS disk)
 add_subdirectory_ifdef(CONFIG_DSP dsp)
 add_subdirectory_ifdef(CONFIG_EMUL emul)
-add_subdirectory_ifdef(CONFIG_FIDO2 fido2)
 add_subdirectory_ifdef(CONFIG_IMG_MANAGER dfu)
 add_subdirectory_ifdef(CONFIG_INPUT input)
 add_subdirectory_ifdef(CONFIG_JWT jwt)
