+++ subsys/authentication/fido2/fido2_crypto.c
